Fix recurring chores
Recurring chores were not being added to the apartment
correctly because chores added in quick succession would
access a stale apartment cache and overwrite each other.

For instance: if the apartment held [chore1] and
chore2 and chore3 wanted to be added. Both chores
would access the apartment and get [chore1]. When
chore2 was added, the apartment would update to
[chore1, chore2]. When chore3 was added, the apartment
would update to [chore1, chore3].

To fix this, looping through chores that needed to be
added had a callback to ensure that the apartment cache
would not be stale when attempting to add chores
in quick succession.